Add structs for MusicBrainz API entities
Structs for the Artist, Work, and Recording structs. In the intended
workflow, a query is sent to the MusicBrainz database with an artist and
song title. This should result in multiple Recordings, each of which
links to one or more Artists, the Work (the composition), and one or
more Releases (albums containing the recording). For example, covers,
samples, and derivations all link to the same Work, which is key for
discovering relationships between Works, which will be the primary level
of analysis.

Importantly, all these data models store the MusicBrainz ID (mbid) for
future API requests. Recordings also store the mbid to the Work it is
linked to.

If an analysis is desired for the Release level, then an appropriate
data model may be defined.